ktr_platform/flows/b660344d-5c31-4e81-9191-2a2bd5bfba96/src/transformers/values3_2_2_2_transformer.py
"""
Transformação de dados - values3 2 2 2
Gerado automaticamente do KTR: importatabelas.ktr
"""
import pandas as pd
import numpy as np
from loguru import logger
from typing import Dict, Any, Optional, List
from ..utils.validation_utils import ValidationUtils

class Values3222PipelineTransformer:
    """
    Transformer para: values3 2 2 2
    Tipo: SelectValues
    """
    
    def __init__(self, config: Optional[Dict[str, Any]] = None):
        self.config = config or {}
        self.validator = ValidationUtils()
        
    def transform(self, df: pd.DataFrame) -> pd.DataFrame:
        """
        Executa as transformações nos dados
        """
        logger.info(f"🔄 Iniciando transformação: values3 2 2 2")
        
        try:
            if df.empty:
                logger.warning("⚠️ DataFrame vazio recebido para transformação")
                return df
            
            original_count = len(df)
            
            # Transformação genérica
            
            # Adicionar metadata de transformação
            df['transformed_at'] = pd.Timestamp.now()
            df['transformation_applied'] = 'values3 2 2 2'
            
            final_count = len(df)
            logger.info(f"✅ Transformação concluída: {original_count} → {final_count} registros")
            
            return df
            
        except Exception as e:
            logger.error(f"❌ Erro na transformação values3 2 2 2: {e}")
            raise
    
    def validate_input(self, df: pd.DataFrame) -> bool:
        """
        Valida se os dados de entrada estão corretos
        """
        return self.validator.validate_dataframe(df)
    
    def get_required_columns(self) -> List[str]:
        """
        Retorna as colunas obrigatórias para esta transformação
        """
        return []  # Definir baseado no tipo de transformação
