ktr_platform/flows/b660344d-5c31-4e81-9191-2a2bd5bfba96/requirements.txt
loguru>=0.7.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
sqlalchemy>=2.0.0